validator filters to own context
