+++ src/contexts/UI/index.tsx
@@ -10,7 +10,6 @@ import { MaybeAccount } from 'types';
 import { useConnect } from '../Connect';
 import { useNetworkMetrics } from '../Network';
 import { useStaking } from '../Staking';
-import { useValidators } from '../Validators';
 import { useBalances } from '../Balances';
 import { useApi } from '../Api';
 import { defaultUIContext } from './defaults';
@@ -22,11 +21,9 @@ export const UIContext =
 export const useUi = () => React.useContext(UIContext);
 
 export const UIProvider = ({ children }: { children: React.ReactNode }) => {
-  const { isReady, consts, network } = useApi();
+  const { isReady, network } = useApi();
   const { accounts: connectAccounts, activeAccount } = useConnect();
   const { staking, eraStakers, inSetup } = useStaking();
-  const { meta, session } = useValidators();
-  const { maxNominatorRewardedPerValidator } = consts;
   const { metrics } = useNetworkMetrics();
   const { accounts } = useBalances();
 
@@ -80,12 +77,6 @@ export const UIProvider = ({ children }: { children: React.ReactNode }) => {
   // services
   const [services, setServices] = useState(getAvailableServices());
   const servicesRef = useRef(services);
-
-  // validator filtering
-  const [validatorFilters, setValidatorFilters]: any = useState([]);
-
-  // validator ordering
-  const [validatorOrder, setValidatorOrder]: any = useState('default');
 
   // staking setup persist
   const [setup, setSetup]: any = useState([]);
@@ -163,178 +154,6 @@ export const UIProvider = ({ children }: { children: React.ReactNode }) => {
 
   const setSideMenu = (v: number) => {
     setSideMenuOpen(v);
-  };
-
-  const setValidatorsOrder = (by: string) => {
-    setValidatorOrder(by);
-  };
-
-  const setValidatorsFilter = (filter: any) => {
-    setValidatorFilters(filter);
-  };
-
-  // Validator list filtering functions
-
-  const toggleAllValidatorFilters = (toggle: number) => {
-    if (toggle) {
-      setValidatorsFilter([
-        'all_commission',
-        'blocked_nominations',
-        'over_subscribed',
-        'missing_identity',
-        'inactive',
-      ]);
-    } else {
-      setValidatorFilters([]);
-    }
-  };
-
-  const toggleFilterValidators = (f: string) => {
-    const filter = [...validatorFilters];
-    const action = filter.includes(f) ? 'remove' : 'push';
-
-    if (action === 'remove') {
-      const index = filter.indexOf(f);
-      filter.splice(index, 1);
-    } else {
-      filter.push(f);
-    }
-    setValidatorsFilter(filter);
-  };
-
-  const applyValidatorFilters = (
-    list: any,
-    batchKey: string,
-    filter: any = validatorFilters
-  ) => {
-    if (filter.includes('all_commission')) {
-      list = filterAllCommission(list);
-    }
-    if (filter.includes('blocked_nominations')) {
-      list = filterBlockedNominations(list);
-    }
-    if (filter.includes('over_subscribed')) {
-      list = filterOverSubscribed(list, batchKey);
-    }
-    if (filter.includes('missing_identity')) {
-      list = filterMissingIdentity(list, batchKey);
-    }
-    if (filter.includes('inactive')) {
-      list = filterInactive(list);
-    }
-    return list;
-  };
-
-  const resetValidatorFilters = () => {
-    setValidatorFilters([]);
-    setValidatorOrder('default');
-  };
-
-  const filterMissingIdentity = (list: any, batchKey: string) => {
-    if (meta[batchKey] === undefined) {
-      return list;
-    }
-    const filteredList: any = [];
-    for (const validator of list) {
-      const addressBatchIndex =
-        meta[batchKey].addresses?.indexOf(validator.address) ?? -1;
-
-      // if we cannot derive data, fallback to include validator in filtered list
-      if (addressBatchIndex === -1) {
-        filteredList.push(validator);
-        continue;
-      }
-
-      const identities = meta[batchKey]?.identities ?? [];
-      const supers = meta[batchKey]?.supers ?? [];
-
-      // push validator if sync has not completed
-      if (!identities.length || !supers.length) {
-        filteredList.push(validator);
-      }
-
-      const identityExists = identities[addressBatchIndex] ?? null;
-      const superExists = supers[addressBatchIndex] ?? null;
-
-      // validator included if identity or super identity has been set
-      if (identityExists !== null || superExists !== null) {
-        filteredList.push(validator);
-        continue;
-      }
-    }
-    return filteredList;
-  };
-
-  const filterOverSubscribed = (list: any, batchKey: string) => {
-    if (meta[batchKey] === undefined) {
-      return list;
-    }
-    const filteredList: any = [];
-    for (const validator of list) {
-      const addressBatchIndex =
-        meta[batchKey].addresses?.indexOf(validator.address) ?? -1;
-
-      // if we cannot derive data, fallback to include validator in filtered list
-      if (addressBatchIndex === -1) {
-        filteredList.push(validator);
-        continue;
-      }
-      const stake = meta[batchKey]?.stake ?? false;
-      if (!stake) {
-        filteredList.push(validator);
-        continue;
-      }
-      const totalNominations = stake[addressBatchIndex].total_nominations ?? 0;
-      if (totalNominations < maxNominatorRewardedPerValidator) {
-        filteredList.push(validator);
-        continue;
-      }
-    }
-    return filteredList;
-  };
-
-  const filterAllCommission = (list: any) => {
-    list = list.filter(
-      (validator: any) => validator?.prefs?.commission !== 100
-    );
-    return list;
-  };
-
-  const filterBlockedNominations = (list: any) => {
-    list = list.filter((validator: any) => validator?.prefs?.blocked !== true);
-    return list;
-  };
-
-  const filterInactive = (list: any) => {
-    // if list has not yet been populated, return original list
-    if (session.list.length === 0) {
-      return list;
-    }
-    list = list.filter((validator: any) =>
-      session.list.includes(validator.address)
-    );
-    return list;
-  };
-
-  // Validator list ordering functions
-
-  const orderValidators = (by: string) => {
-    const order = validatorOrder === by ? 'default' : by;
-    setValidatorsOrder(order);
-  };
-
-  const applyValidatorOrder = (list: any, order: string) => {
-    if (order === 'commission') {
-      return orderLowestCommission(list);
-    }
-    return list;
-  };
-
-  const orderLowestCommission = (list: any) => {
-    const orderedList = [...list].sort(
-      (a: any, b: any) => a.prefs.commission - b.prefs.commission
-    );
-    return orderedList;
   };
 
   // Setup helper functions
@@ -481,12 +300,6 @@ export const UIProvider = ({ children }: { children: React.ReactNode }) => {
       value={{
         setSideMenu,
         setUserSideMenuMinimised,
-        orderValidators,
-        applyValidatorOrder,
-        applyValidatorFilters,
-        resetValidatorFilters,
-        toggleFilterValidators,
-        toggleAllValidatorFilters,
         toggleService,
         getSetupProgress,
         getSetupProgressPercent,
@@ -497,8 +310,6 @@ export const UIProvider = ({ children }: { children: React.ReactNode }) => {
         sideMenuOpen,
         userSideMenuMinimised: userSideMenuMinimisedRef.current,
         sideMenuMinimised,
-        validatorFilters,
-        validatorOrder,
         services: servicesRef.current,
         onSetup,
         isSyncing,
